Special case colors in pattern relaxation.
We treat colors as a special case.
In a complex object (e.g. a dog), an object and a large
number of its sub-components will share the same color.
Usually if the color is irrelevant to a pattern in one place,
it is irrelevant in many other places.
 Relaxing the pattern by peeling the color predicates off
one by one can be very slow,
 especially since such large patterns are slow to match in
the first place.
 Therefore, if we remove a color predicate in one place,
we remove it
 from the entire pattern.