Replace equals-signs with at-signs because HTCondor doesn't like them
I am not sure why this is, but it seems to be true. SLURM is fine with
them in the job name, and if Pegasus had a problem with them I would
expect it to fail with some Condor-unrelated error, so probably it is
Condor that is failing.

(Unless Condor provides an escape that Pegasus is mistakenly not using
or something. I don't know.)