Fix cross-situational object learner freeze
Previously, the cross-situational object learner would freeze the first time it
observed and then tried to describe a dog. It froze attempting to find a
second match for the dog pattern.

This fix changes the cross-situational object learner to be like other object
learners and only look for a single match.