Add argument to control number of samples per object
This is required for the script to behave properly on both train and test
splits, which have differing numbers of samples. Previously we didn't have to
worry about this because we were dropping situations where we couldn't find the
strokes -- which meant we'd find a bunch of files "missing" for test and skip
over those nonexistent examples.