Fix crash when learner hypothesis and perception have continuous nodes
The crash happened because we were actually updating the new hypothesis created
from the current perception, rather than updating the old hypothesis.

The core fix is the change to perception_graph. Because of reordering the graph
and pattern we also have to reorder things in clients' allowed_matches args and
in the template intersection code. This is necessary for non-object learners to
learn correctly as otherwise they raise key errors or fail to learn due to
incorrect match postprocessing.