GNN decode: Fix how single-object output is saved
In single-object mode, we treat any situation with multiple objects as having
just one object. However, our prediction code didn't account for this. Because
of that, if stroke extraction detected >1 object we would only overwrite the
label (concept_name) for the first object listed. And that was a problem because
stroke extraction was producing, for each object detected, a default label
(concept_name) equal to the ground truth label for the situation. Thus, every
label but the first (the GNN's) was guaranteed to be correct.